% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{connections}
\alias{connections}
\title{Connection information needed to draw segments between face tracking points.}
\format{An object of class \code{data.frame} with 63 rows and 2 columns.}
\usage{
connections
}
\description{
Connection information needed to draw segments between face tracking points.
}
\author{
Jan Niklas Schneider \email{jan_schneider@live.de}
}
\keyword{datasets}
