% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{loadings}
\alias{loadings}
\title{Print loadings of plsr object}
\usage{
loadings(x, mat = NULL)
}
\arguments{
\item{x}{A plsr object.}

\item{mat}{Which matrix to print. Can be "V" or "U", if NULL (default) will print both.}
}
\description{
This will print the loading matrices V and U that project from original data spaces X and Y to latent space.
}
\examples{
plsr_obj = pls(rating_data,tracking_data,10,10)
loadings(plsr_obj) #show V and U
loadings(plsr_obj,"V") #show V only
loadings(plsr_obj,"U") #show U only
}
