% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{plot.plsr}
\alias{plot.plsr}
\title{Plot function for plsr objects}
\usage{
\method{plot}{plsr}(x, ...)
}
\arguments{
\item{x}{The plsr object.}

\item{...}{Further arguments.}
}
\description{
Plots information about a plsr object. The following plots will be generated:
\itemize{
\item barplot of p-values of latent variables estimated via permuatation testing
\item Histograms of the distributions of latent variables derived via permutation testing
\item A plot showing the effect of the first latent variable on the original data spaces
\item Several plots to visualize bootstrapping results
}
}
\examples{
\donttest{
plsr_obj = pls(rating_data,tracking_data)
plot(plsr_obj)
}

}
