% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shiny.R
\name{plot_default}
\alias{plot_default}
\title{Default plot function for plsr shiny app}
\usage{
plot_default(x, time_steps = 10, t)
}
\arguments{
\item{x}{A vector of predicted data to plot}

\item{time_steps}{Number of time steps}

\item{t}{Which time step to plot}
}
\description{
Default plot function for plsr shiny app
}
