% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{plot_latent_variables}
\alias{plot_latent_variables}
\title{Plots latent variables}
\usage{
plot_latent_variables(plsr_obj, lv_num = 1, sd = 3, frame = 1)
}
\arguments{
\item{plsr_obj}{A plsr object}

\item{lv_num}{An integer or list of integer specifying which latent variables to plot}

\item{sd}{Range in standard deviations from +[sd] to -[sd]}

\item{frame}{Which timestep to plot}
}
\description{
Plots latent variables
}
