% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{pls}
\alias{pls}
\title{Run partial least square analysis}
\usage{
pls(X, Y, n_perm = 1000, n_boot = 1000, scale = T, verbose = F)
}
\arguments{
\item{X}{A matrix of m observations on n dimensions}

\item{Y}{A matrix of m observations on n dimensions}
}
\value{
A PLS Object
}
\description{
Run partial least square analysis
}
