% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{print.plsr}
\alias{print.plsr}
\title{Print plsr object}
\usage{
\method{print}{plsr}(x, ...)
}
\arguments{
\item{x}{A plsr object.}

\item{...}{Further arguments.}
}
\description{
Prints information about a plsr object.
}
