% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{rating_data_emo_means}
\alias{rating_data_emo_means}
\title{Mean emotion ratings on the dimensions happy, sad, surprised, disgusted, angry, fearful and interested}
\format{An object of class \code{data.frame} with 40 rows and 7 columns.}
\usage{
rating_data_emo_means
}
\description{
Mean emotion ratings on the dimensions happy, sad, surprised, disgusted, angry, fearful and interested
}
\author{
Jan Niklas Schneider \email{jan_schneider@live.de}
}
\keyword{datasets}
