% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{summary.plsr}
\alias{summary.plsr}
\title{Summary of plsr object}
\usage{
\method{summary}{plsr}(object, ...)
}
\arguments{
\item{object}{A plsr object.}

\item{...}{Further arguments.}
}
\description{
Summary of plsr object
}
\examples{
\donttest{
plsr_obj = pls(rating_data,tracking_data)
summary(plsr_obj)
}
}
