% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tracking_data_emo_means}
\alias{tracking_data_emo_means}
\title{Face tracking data on 40 videos of emotional facial expressions.}
\format{An object of class \code{data.frame} with 40 rows and 1360 columns.}
\usage{
tracking_data_emo_means
}
\description{
Face tracking data on 40 videos of emotional facial expressions.
}
\author{
Jan Niklas Schneider \email{jan_schneider@live.de}
}
\keyword{datasets}
