% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{explained_variance}
\alias{explained_variance}
\title{Calculates explained variance per component of the original data sets X and Y}
\usage{
explained_variance(plsr_obj)
}
\arguments{
\item{plsr_obj}{A plsr object}
}
\description{
Calculates explained variance per component of the original data sets X and Y
}
