% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{loadings.plsr}
\alias{loadings.plsr}
\title{Print loadings of plsr object}
\usage{
loadings.plsr(x, mat = NULL)
}
\arguments{
\item{x}{A plsr object}

\item{mat}{Which matrix to print (U or V), if NULL (default) will print both}
}
\description{
This will print the loading matrices V and U that project from original data spaces X and Y to latent space.
}
\examples{
\dontrun{
loadings(x)
loadings(x,"U")
loadings(x,"V")
}
}
