% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{new_plsr}
\alias{new_plsr}
\title{Constructor for plsr objects}
\usage{
new_plsr(decomp = list(), perm = list(), bootstrp = list(),
  sclng = list(), org_dat = list(), cl = list())
}
\arguments{
\item{decomp}{List of singular value decomposition results}

\item{perm}{List of permutation testing results}

\item{bootstrp}{List of bootstrapping results}

\item{sclng}{List of scaling paramters applied to original data}

\item{org_dat}{List of original data}

\item{cl}{Call of pls function}
}
\description{
Constructor for plsr objects
}
\examples{
#Creating an empty plsr object
d=p=b=s=o=c = list()
plsr_obj=new_plsr(decomp=d,perm=p,bootstrp=b,sclng=s,org_dat=o,cl=c)
}
