% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{permutation_precision}
\alias{permutation_precision}
\title{Calculates the precision of the p-value estimated by permutation testing}
\usage{
permutation_precision(p, k)
}
\arguments{
\item{p}{The p value}

\item{k}{Number of permutation iterations}
}
\description{
Calculates the precision of the p-value estimated by permutation testing
}
