% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_func.R
\name{pls}
\alias{pls}
\title{Run partial least squares analysis}
\usage{
pls(X, Y, n_perm = 10, n_boot = 10, scale = T, verbose = F,
  alpha = 0.05)
}
\arguments{
\item{X}{A matrix of m observations on n dimensions}

\item{Y}{A matrix of m observations on n dimensions}

\item{n_perm}{Number of permutation iterations}

\item{n_boot}{Number of bootstrap iterations}

\item{scale}{Scaling of X and Y (Boolean)}

\item{verbose}{Provides additional output}

\item{alpha}{The significance level}
}
\value{
A PLS Object
}
\description{
Run partial least squares analysis
}
